compiler: implement comparision operators
code_gen: binop: fixed a bug where the right hand expression was being
overridden by the left hand side while popping the left hand side for
final operation. Now this is handled by popping the left hand into `rcx`
and then exchanging it with `rax` so that the result of the operation
always ends up in `rax`.