convert codebase to cpp
Why? Because I needed the buf_printf() to print to string
too. But C lacks the ability to define functions like printf().
So when printing type names in cg_llvmir.c, there is no way to redirect
it to a file for the llvm static compiler to compile it to assembly.