Add tests and run them internally
If the `test` target is run in the Makefile, it now generates a new
executable named `test` in the `build` directory, which essentially
replaces the `src/main.c` file with `src/tests/tests.c` file and runs
the newly-generated executable which tests all parts of the compiler
pipeline. This replaces the `test.sh` file.

Note that for tests like `no input files` or any test which uses stuff
externally to the compiler (ile io, temp files), we still need to use
the `test.sh` file.