parse: Change array literal syntax to `[]`
Currently there is no way to specify the type of the literal. We could
change the syntax by prefixing the literal with a type, but that will
cause arbitrary look-ahead or syntax ambiguity complications. The other
option is to append the type at the end, specifically the child type of the
array, as the compiler can infer the number of elements.