sema: Implement `return` and add `noreturn` type
msg: Fixed a bug wherein the function `print_source_line` was
erroneously printing characters after the span before the newline even
though there was no additional characters after it. Turns out the
initial counter of the for loop was mistakenly set to the multi-line
EOS (end of span), instead of the newline of the first line.