sema, cg: check `noreturn` in last func statement
We now check if the last statement of a function body is of type
`noreturn` (specifically the expression statement).

In sema: for non-void (and non-noreturn) functions, we check if the last
statement is a noreturn, and if it is not, we emit an "non-void
function does not return a value" error.

In cg: for void and noreturn functions, we have to implicitly emit a
`ret void` or an `unreachable` if the user did not `return;` or call a
noreturn function. So we check if the last statement is a noreturn, and
it is not, we emit one. LLVM doesn't like if the terminator is
duplicated, so we have to have this check.